Add files via upload
This is my internship project (Solar Secure Solutions) where I have developed a E-Commerce website for food industry or you can call it a website to buy food items from.  
Along with frontend technologies, backend elements like node.js , express.js and EJS has been used in the project. 
This is just a small start to my footstep in the IT world.! . Many more to go.